Resolved bug where IDs weren't loading in from localstorage API.
Conditions:
When constructing tasks in projects object from localstorage data, the IDs
of the individual task was not persisting. This led to some large ui breakdowns
as the id is a critical component to layout.

Cause:
When committed to json, the task objects are no longer passed by reference, each becomes its own object. SO changes no longer reflect the other. The constructor conditions to build task didn't account for id as I would let the id function reset all ids. So the id would be undefined then it would stringify back to json, completely losing that property in the process.

Solution:
I set the constructor to map the id property as well.

Concerns:
Now that these objects are no longer sharing a reference. When the ids reset and change due to some
task being deleted and removed, the ids will not line up for task that are purpotedly the same.
*Possible solution* => implement id generator so changing of ids is not necessary.